Minor style fixes
JerryScript-DCO-1.0-Signed-off-by: László Langó llango.u-szeged@partner.samsung.com

+++ src/example5.c
@@ -148,5 +148,5 @@ main (int argc, char * argv[])
   // Cleanup engine
   jerry_cleanup ();
 
-  return status;
+  return (int) status;
 }
